Added frame output in max
Each Frame of data is now recorded to a file in max. The frame rate has
been set 1 in order to not lock up the system while trying to do this
operation. Also added the beginnings of a Swift command line tool for
combining the individual frames into a single csv file.